Membuat class Family
