+++ src/lib/Employee.java
@@ -1,8 +1,6 @@
 package lib;
 
 import java.time.LocalDate;
-import java.util.LinkedList;
-import java.util.List;
 
 public class Employee {
     private Person person;
@@ -13,12 +11,6 @@ public class Employee {
     private int otherMonthlyIncome;
     private int annualDeductible;
 
-    // Data Keluarga
-    private String spouseName;
-    private String spouseIdNumber;
-    private List<String> childNames;
-    private List<String> childIdNumbers;
-
     public Employee(Person person, LocalDate dateJoined, int monthWorkingInYear, int monthlySalary, int otherMonthlyIncome, int annualDeductible) {
         this.person = person;
         this.dateJoined = dateJoined;
@@ -26,9 +18,6 @@ public class Employee {
         this.monthlySalary = monthlySalary;
         this.otherMonthlyIncome = otherMonthlyIncome;
         this.annualDeductible = annualDeductible;
-
-        childNames = new LinkedList<String> ();
-        childIdNumbers = new LinkedList<String> ();
     }
 
     public void setMonthlySalary(int grade) {
@@ -58,16 +47,6 @@ public class Employee {
         this.otherMonthlyIncome = income;
     }
 
-    public void setSpouse(String spouseName, String spouseIdNumber) {
-        this.spouseName = spouseName;
-        this.spouseIdNumber = spouseIdNumber;
-    }
-
-    public void addChild(String childName, String childIdNumber) {
-        childNames.add(childName);
-        childIdNumbers.add(childIdNumber);
-    }
-
     public int getAnnualIncomeTax() {
 
         //Menghitung berapa lama pegawai bekerja dalam setahun ini, jika pegawai sudah bekerja dari tahun sebelumnya maka otomatis dianggap 12 bulan.
@@ -79,6 +58,6 @@ public class Employee {
             monthWorkingInYear = 12;
         }
 
-        return TaxFunction.calculateTax(monthlySalary, otherMonthlyIncome, monthWorkingInYear, annualDeductible, spouseIdNumber.equals(""), childIdNumbers.size());
+        return TaxFunction.calculateTax(monthlySalary, otherMonthlyIncome, monthWorkingInYear, annualDeductible, person.getSpouseName().equals(""), person.getChildIdNumbers().size());
     }
 }
